moved shiftChar function to separate file
